pip-requirements.txt
beautifulsoup4==4.8.2
html5lib==1.0.1
requests==2.22.0
pyOpenSSL==19.1.0
pyyaml==5.3

